Pair programming review
